DotNET: Replace nethost search with static linked library. (#1869)
* Replace nethost search with static linked library. Update directory structure to match external convention.

* Don't show assert message on failed resolve of get_hostfxr_path.
